✨ Feat: Create entry for past days automatically
If there is a media file in .tmp but no entry for day the media file was
created on, then this day entry is created with no text, ie. empty
pre-tags but with the proper media tags, ie. img and video.

This solves the problem that some media files weren't added to a diary
entry because there was none. They remained in .tmp. Because I can't
write an entry for each media file, I took in the past, I will create a
diary entry with the media files only and may add a description later.