🚀 Improve: Own function for adding media files
Separated add_media_files() into a function that takes the target html
file and the directory from which the media files should be added. By
doing so, it is possible to add media files "on the fly" of a directory
without inserting them into the diary. This makes sensefor vacations
where I usually take multiple fotos I dont wnat to save twice.